Code refactoring, added TransferBuffer<T> base type
